move controller out of internal and some little fixes
