+++ services/platform/operator/controller/install_controller.go
