+++ services/platform/operator/controller/persistence.go
