+++ services/platform/operator/controller/secrets/generator.go
