+++ services/platform/operator/controller/secrets/generator_test.go
