+++ services/platform/operator/controller/talos/groupversion_info.go
